BUG-58: refactor to take advantage of netty
This patchset reworks the framework to take advantage of features in
netty, shortcutting many of the previous concepts. This lead to a major
refactor, reorganizing the way things are glued together.

SessionProposal/Checkers have been removed in favor of an explicit
negotiation piece, which is put on the channel pipeline to negotiate
session parameters -- uopn success it replaces itself with a complete
session.

ProtocolSessionInboundHandler has been integrated into
AbstractProtocolSession, handling the tasks normally.

Finally, Dispatcher uses all this restructure to implement Session
reconnect in a relatively straightforward manner.

Change-Id: I84a6b4c99282212505e6d91ab40da9867ca95e53
Signed-off-by: Robert Varga <rovarga@cisco.com>