BUG-48: message send should return future
The next step will see us integrate sendMessage() into MD-SAL RPC
service. The RPC's fate has to track the message as it leaves the
system -- they return a Future, which resolves one the RPC is complete.

The definition in this case requires that we allow sendMessage() users
to obtain the Future which will fail if the message does not hit the
socket.

Change-Id: If0e086d5a53b17712dfb05730d73cca68f3242e7
Signed-off-by: Robert Varga <rovarga@cisco.com>