BUG-191: provide a way for executors to provide scheduling feedback
The idea is that the executor is hooked onto the ListenableFuture,
kicking in if it succeeds. Should the insn be cancelled, the future
will be, too.

The notification part is handled internally until the Future succeeds --
once it does, the executor can either start immediately executing, after
checking with the Instruction, or inform it why it delayed execution.

Once execution completes, the executor pushes the appropriate details
into the instruction.

Change-Id: I9dd9c29c876b890cadf842941c1e59e0f394cf84
Signed-off-by: Robert Varga <rovarga@cisco.com>