Rework BGP configuration
This patch reworks the structure of BGP configuration so that it follows
the logical organization of the RIB.

The core change is the creation of BGPPeer as a full-blown configuration
item. This encapsulates the previous proposal configuration and has the
inverse relationship with the RIB: the RIB is its dependency and it
pulls things like the local AS and BGP identifier from it.

Futher improvement is the introduction of bgp-table-type, which allows
explicit listing of tables which should be advertized in the BGP Open
message as well as specification of which tables should be tracked in
the local RIB.

Change-Id: I905759a3d9e780394579238d48ff0aee009a5544
Signed-off-by: Robert Varga <rovarga@cisco.com>
Signed-off-by: Maros Marsalek <mmarsale@cisco.com>
Signed-off-by: Robert Varga <rovarga@cisco.com>