BUG-633: rework packaging
First we split out tcpmd5-api component, which contains the basic
interfaces and classes.

Second we rename tcpmd5-core to tcpmd5-nio to better reflect the fact
this is the binding to java.nio package. We also introduce
DefaultKeyAccessFactoryFactory, which provides a facade to selecting the
best-functioning KeyAccessFactory.

Third we introduce NativeSupportUnavailableException to indicate when
the native support library fails to load/initialize.

Finally we fix invalid javaChannel() reference in tcpmd5-netty, which
should have been called on the superclass.

Change-Id: I1243189ab728abecc473b3aa5a38e9e167b80aad
Signed-off-by: Robert Varga <rovarga@cisco.com>