Update patch set 1
Patch Set 1:

Robert, this is because i release using the default profile, and in that it includes some of the jni modules but not all. So the fix is to make the set of modules in the default profile releasable. I'm not sure what is your intention in performing the release, if to release also the JNI artifacts or no, you guys have that answer. Certainly it would be much better if you place all the artifacts you want to released in the default profile so we don't have to guess on a per-project base it.

Patch-set: 1
Reviewer: Gerrit User 19 <19@7fc14799-209e-464c-9743-7a06c2c21a81>
Label: Verified=0