Update patch set 1
Patch Set 1: Code-Review-1

So the unfortunate thing here is that the JNI stuff compiles only on Linux. So we do have specific requirements on the machine where the release happens (64bit Linux, with TCP-MD5 support enabled). We will not make it the default, as that will break anyone running it everywhere else.

So yes, everything in our project should be released and for the proper profiles which need to be activated please see our release job (jni and javadoc? I don't remember)

Patch-set: 1
Reviewer: Gerrit User 649 <649@7fc14799-209e-464c-9743-7a06c2c21a81>
Label: Code-Review=-1