Update LSPDB incrementally
Now that the datastore can properly handle augment removal and fast
merges, we can get rid of the full rebuild we were doing.

This patch restores the previous logic, such that we enqueue only LSP
updates into the datastore.

The flip of 'synced' boolean is handled as a merge of a single leaf, so
should propagate writhout incurring any other overhead.

Change-Id: I488e15ebef3abdeab6df791b7ed3b7fe2c266330
Signed-off-by: Robert Varga <rovarga@cisco.com>