BUG-1116: Rework comparator and encapsulate transactions
Having the route comparator in the peer was just a hack. It really is
part of the local rib, so this patch moves it to the appropriate place.

Since we are changing interfaces, change them some more by not passing
raw transaction around, but rather a proper transaction context. This
 will allow us to transparently route it to other peers for dissemination
purposes.

Change-Id: Ib6e5362f32dcf9f2582bb76c364c8028a2c08c54
Signed-off-by: Robert Varga <rovarga@cisco.com>
Signed-off-by: Milos Fabian <milfabia@cisco.com>