Support proper route redistribution
For proper route redistribution, we need to properly process received
attributes even if we do not recognize them. The parser needs to output
them as received and on egress we need to munge the value with a new
header, which will force the partial bit to be set to true.

Change-Id: Icc5d3b6393dae7ba6d83df0f8beb596cccc9559a
Signed-off-by: Robert Varga <rovarga@cisco.com>