BUG-2383: introduce OffsetMap
For the sectionprocess we will need a compact way of tracing
peer/attribute mappings. Storing things in array is the most compact
form, but we need to manage a lookup table. This is the role of
OffsetMap, which maps an UnsignedInteger (e.g. a BGP Identifier) to an
offset. We then need to only allocate a single array for each entry to
hold the actual values.

Change-Id: I7c4c8b841cca0330f9646e5e629a824fd929f181
Signed-off-by: Robert Varga <rovarga@cisco.com>