BUG-2383: cache effective attributes in import
This patch introduces a simple identity cache decorator around
AbstractImportPolicy. It is then used by ImportPolicyPeerTracker and
EffectiveRibInWriter to make sure we do not needlessly explode
attribute objects. This would typically happen when we receive multiple
routes within a single message and the import policy modifies the
attributes -- leading to a per-route attribute object.

Change-Id: I8581df3e3a2aae07a16d44f2554026afbad264fd
Signed-off-by: Robert Varga <rovarga@cisco.com>