BUG-3225 : fix NPE
Import policy requested by Effective RIB was yet to be put in policy database,
therefore the NPE occurred. Switched off listening for peer-role changes itself,
the change is now checked in Effective RIB. Since we are listening to a different
subtree, the handling in Effective RIB needed to be changed accordingly.

Change-Id: I0df6cf2234a27d99445af07dd47c9039a0418cb4
Signed-off-by: Dana Kutenicsova <dkutenic@cisco.com>