Incorrect logic in withdraw.
Withdraw is second argument in buildUpdate, not first.

So if you are trying to ask BGP to withdraw a route by putting
an empty route list into "withdrawn" and the route to be withdrawn
into "to advertise" you most likely not going to get very far
(unless there is one more bug somewhere to cancel this one out).

Change-Id: I71604300b485806ffe37dbe883abf5848121a2b5
Signed-off-by: Anton Ivanov <aivanov@brocade.com>