Incorrect logic in withdraw.
Withdraw is second argument in buildUpdate, not first.

So if you are trying to ask BGP to withdraw a route by putting
an empty route list into "withdrawn" and the route to be withdrawn
into "to advertise" you most likely not going to get very far
(unless there is one more bug somewhere to cancel this one out).

Same applies for advertise.

Change-Id: I6d90ba078b9008ff38c79d1d3b12b3dccbe01716
Signed-off-by: Anton Ivanov <aivanov@brocade.com>