Incorrect logic in withdraw.
Withdraw is second argument in buildUpdate, not first.

So if you are trying to ask BGP to withdraw a route by putting
an empty route list into "withdrawn" and the route to be withdrawn
into "to advertise" you most likely not going to get very far
(unless there is one more bug somewhere to cancel this one out).

The same is valid for advertise - they were swapped.

Change-Id: I2b5488922824a7ac0c1135e85626927bc354467e
Signed-off-by: Anton Ivanov <aivanov@brocade.com>