BUG-3888 : fix comparing ASnumbers
In order to allow BGP connection, expected remote AS
(configured in ODL) needs to match AS number received
in peer Open message. Added expected remote AS to BGP
Session Preferences to allow for this comparison.

Moved BGPClientSessionValidator code to StrictBGPPeerRegistry
so we do the same checks on inbound and outbound BGP
sessions (e.g. checks for same BGP ID and for 4-byte AS
capability).   Then deleted BGPSessionValidator and both the
BGPClientSessionValidator and BGPSessionValidator.

Merged Client and Server negotiators and changed to use inbound
flag in getPeer etc. to clarify whether source of BGP session
is local.

Change-Id: I0be64c30980b47a6b9bd3ba7263e12b98af68524
Signed-off-by: Dana Kutenicsova <dkutenic@cisco.com>
Signed-off-by: Giles Heron <giheron@cisco.com>