BUG-4070: Define new specific Peer role for App Peer.
Currently App Peer is defined a iBGP, therefore it doesnt
advertise to iBGP peers. Setting it as an RR, means that
it will advertise ORIGINATOR_ID and CLUSTER_ID, which
is undesirable.
Instead we create a new role for App Peer, INTERNAL.
Which will act as an RR-client but without advertise
undesirable fields ORIGINATOR_ID and CLUSTER_ID.

Change-Id: I1385158806838623824773c3031556fefa2a0262
Signed-off-by: Claudio D. Gasparini <cgaspari@cisco.com>