Update patch set 2
Patch Set 2: Code-Review-1

This is not what needs to happen in the scope of the bug. We need to make sure we have a new peer role, Internal or similar, which is used only for the application peer. Then we need to export the routes to all peers (except originator) if they come from an Internal peer.

Patch-set: 2
Label: Code-Review=-1