BUG-4197: BGP OpenConfig SPI
Defines the APIs to track a BGP configuration modules status changes and
BGP OpenConfig data writers consumed as a service injected by config subsystem into
BGP configuration module.
The goal is to react on changes in the BGP configuration module instances - the configuration
modules are mapped to the OpenConfig BGP data model instances and kept in a datastore.

Change-Id: Ic38d5049d4d7e2f51f05a0e44e28626fc48e7ab7
Signed-off-by: Milos Fabian <milfabia@cisco.com>