BUG-3771: Use sync on FutureChannels intead of await
"Await" waits until future has finished and throw execution
if thread was interrupted. "Sync"" instead waits for this
future until it is done, and re throws the cause of
the failure if this future failed.
When closing the module, first we ensure that channel
has been closed and if it didn't fail we proceed
with next closures.

Change-Id: I2d1d6c8c82c3fb27205ee2272fbc559c255da0bd
Signed-off-by: Claudio D. Gasparini <cgaspari@cisco.com>