Use sync on FutureChannels intead of await
Await waits until future has finished and throw exection
if thread was interrupted. Sync instead waits for this
future until it is done, and rethrows the cause of
the failure if this future failed.

Channel is been closed on 2 differents places
Fix it by close channel only once on Topology provider.

Change-Id: I2d1d6c8c82c3fb27205ee2272fbc559c255da0bd
Signed-off-by: Claudio D. Gasparini <cgaspari@cisco.com>