Update patch set 8
Patch Set 8:

Good news for this patch, bad news for testing on Sandbox. Vpnservice job got the same session failure (which does not happen on main Releng Jenkins) when run against master distribution build (on Sandbox).

I am going to run a version of that suite with a sleep somewhere, to hopefully avoid the failure.

Patch-set: 8