BUG-4689: Advertising performance fix II
When AdjRibOutListener get a notification with too many changes,
it can fill output buffer with wrtitten Update messages,
without a chance to flush a socket, because the AdjRibOutListener
is blocked when channel is not writable - deadlock situation until a
savior (KA message) will not be send.
Flushing with every write operation is uneffective, hence try to call
the "flush" in a situation when the channel is unwritable and we want to write
abother Update to the channel.

Change-Id: Ie50b9154d1166d55308c3438c12b9c1b9d3ce19f
Signed-off-by: Milos Fabian <milfabia@cisco.com>