Bug 5558 - Closing session after a large RIB is sent kills BGPCEP
Fixed peer session recovery, in a case of failed transaction chain.
The AdjRibInWriter need working chain when cleaning, also it needs
to be started with fresh transaction chain when session goes up again.
Hence recreate AdjRibInWriter with a new tx chain, when the current
failed.

Fixed duplicated session avoidance. The peer's session should not be
removed from registry if it was not created during this negotiation process.
Otherwise, the active peer's sesion can be removed from registry, which can
later lead to the duplicated session establishment.

Signed-off-by: Milos Fabian <milfabia@cisco.com>
Change-Id: I46ef33d49a11db8465a719af5ba4e1716da14602