Bug 5689: Unhandled message is causing failure
A write transaction is created when message is recieved,
however in case the message is not handled
(i.e. PcError Message without SRP), the transaction
is not sumbitted, which casuse a race with the next transaction.

Submitting the empty transaction of the unhandled message is
a safe way to avoid the problem.

Change-Id: Ie191eec2b9517078f39c5f065243c87b04ddc189
Signed-off-by: Milos Fabian <milfabia@cisco.com>
(cherry picked from commit 9b6b0df9ccad25843a4b9e970fcf134028fb426e)