Update patch set 1
Patch Set 1:

> -the validation should not be done at application layer (BGPPeer),
 > but in the codecs

By codec, you mean BGPUpdateMessageParser.java in this case, right?

 > -for a case of user input (application peer use-case) the mandatory
 > attribute should be enforced at the level of the yang data model
 > (mandatory statement)

The logic to decide whether a particular mandatory attribute should be present is not static but depends on other values E.g. ORIGIN and AS_PATH are mandatory only if NLRI or MP_REACH_NLRI is present. So in case of user input also, we may need to do the check in code?
 > -in case of BGP Multiprotocol (most of the cases) a bit different
 > behavior is defined regarding the next hop attribute/field -
 > https://tools.ietf.org/html/rfc4760#section-3

Yes I had looked at it and included that in logic of method checkMandatoryAttributesPresence(). Anything specific which I am missing?

Patch-set: 1
Label: Verified=0