BUG-5966: Remove received path Id from routeKey
When we map received routes on locRibwriter is required to remove
received pathId to identify the key value (prefix, key-value)as a key.
Otherwise is identifying the combination  of key+path id
as a new route and not as a new path.
This step is only required on extension which support Multiple Path

Change-Id: Ia4d6e86cc187a5db71ecfb66c745b9a2420732da
Signed-off-by: Claudio D. Gasparini <cgaspari@cisco.com>