Add blueprint XML for bgp-parser-spi
Added blueprint XML file, bgp-parser-spi.xml, to the spi bundle that Utilizes
the new blueprint extension, "specific-reference-list" (added by
https://git.opendaylight.org/gerrit/#/c/40267/), to obtain the
specific list of RSVPExtensionProviderActivator OSGi service
instances. The specific expected services are announced via
META-INF/services/org.opendaylight.protocol.bgp.parser.spi.BGPExtensionProviderActivator
resources in the provider bundles which the "specific-reference-list"
extension scans so it knows which services to look for. Once all
expectedservices are obtained, the RSVPExtensionProviderActivator instance
List bean is created and injected into the SimpleRSVPExtensionProviderContextActivator
bean instance which calls start on each.

The SimpleBGPExtensionProviderContextModule was deprecated but still
remains for now to provide the BGPExtensionProviderContext instance
created via blueprint to the config system so it can be injected into
other users. Once the other user(s) are converted to blueprint then the
config yang and Module classes can be removed.

Change-Id: Id41d410b449699ac3d82b3d0352ea84eb66f22ca
Signed-off-by: Tom Pantelis <tpanteli@brocade.com>