Add blueprint XML for bgp-rib-spi
Added blueprint XML file, bgp-rib-spi.xml, to the rib-spi bundle that
Utilizes the new blueprint extension, "specific-reference-list" (added by
https://git.opendaylight.org/gerrit/#/c/40267/), to obtain the
specific list of RIBExtensionProviderActivator OSGi service instances.
The specific expected services are announced via
META-INF/services/org.opendaylight.protocol.bgp.rib.spi.RIBExtensionProviderActivator
resources in the provider bundles which the "specific-reference-list"
extension scans so it knows which services to look for. Once all
expected services are obtained, the RIBExtensionProviderActivator
instance List bean is created and injected into the
SimpleRIBExtensionProviderContextActivator bean instance which calls start on each.

The SimpleRIBExtensionProviderContextModule was deprecated but still
remains for now to provide the RIBExtensionProviderContext instance
created via blueprint to the config system so it can be injected
into other users. Once the other user(s) are converted to blueprint then
the config yang and Module classes can be removed.

Change-Id: I2e3698ad08f021191c9c66ee4e27d55d7bd5e24a
Signed-off-by: Tom Pantelis <tpanteli@brocade.com>