Add blueprint XML wiring for the BGPDispatcher
Added blueprint wiring for the BGPDispatcher. The BGPDispatcherImplModule was
deprecated but still remains for now to provide the RIBExtensionProviderContext
instance created via blueprint to the config system so it can be injected
into other users. Once the other user(s) are converted to blueprint then
the config yang and Module classes can be removed.

Change-Id: Ie1f268a4ee593efeda677e10009b4a351acda17c
Signed-off-by: Tom Pantelis <tpanteli@brocade.com>