Convert bmp extenstions to blueprint
Added blueprint XML file, bmp-spi.xml, to the spi bundle that Utilizes
the new blueprint extension, "specific-reference-list" (added by
https://git.opendaylight.org/gerrit/#/c/40267/), to obtain the
specific list of BmpExtensionProviderActivator OSGi service instances.
The specific expected services are announced via
META-INF/services/org.opendaylight.protocol.bmp.spi.registry.BmpExtensionProviderActivator
resources in the provider bundles which the "specific-reference-list"
extension scan so it knows which services to look for. Once all expected
services are obtained, the BmpExtensionProviderActivator instance List
bean is created and injected into the SimpleBmpExtensionProviderContextActivator
bean instance which calls start on each.

The only BmpExtensionProviderActivator instance wired in 32-bmp.xml is
the BmpActivator in bmp-impl. This instance is now created via the
blueprint XML, bgp-bmp.xml, and the config yang and associated Module
class were removed.

The SimpleBmpExtensionProviderContextModule was deprecated but still
remains for now to provide the BmpExtensionProviderContext instance
created via blueprint to the config system so it can be injected into
other users. Once the other user(s) are converted to blueprint then the
config yang, Module classes can be removed.

Change-Id: I96205d078d9a7ec5fa58eae7c0f66a88227f7bc5
Signed-off-by: Tom Pantelis <tpanteli@brocade.com>