Add blueprint XML wiring for the BGPPeerRegistry
Added blueprint wiring for the StrictBGPPeerRegistry. The
StrictBgpPeerRegistryModule was deprecated but still remains for now
to provide the instance created via blueprint to the config system so it can be
injected into other users. Once the other user(s) are converted to blueprint then
the config yang and Module classes can be removed.

Change-Id: Ieaec3d6e29b457192d3a3acdc5b9e72559403805
Signed-off-by: Tom Pantelis <tpanteli@brocade.com>