Enable restart of CSS modules on blueprint container restart
For the CSS modules whose services are now created via blueprint, if
the corresponding blueprint container is restarted, eg due to config
change, we want to also restart the CSS module. To accomplish this, we
need to associate the blueprint service with the corresponding module
by adding well-known service properties that identify the CSS module's
namespace, module type name, and module instance name. Thus when the
BlueprintContainerRestartServiceImpl encounters the service properties
it will restart the specified CSS module..

Change-Id: Ia1acad4d44e3dff1f3f34bc1f56b33f16862ba62
Signed-off-by: Tom Pantelis <tpanteli@brocade.com>