Add blueprint XML wiring for the BmpDispatcher
Added blueprint wiring for the BGPDispatcher. The BmpDispatcherImplModule
was deprecated but still remains for now to provide the BmpDispatcher
instance created via blueprint to the config system so it can be
injected into other users. Once the other user(s) are converted to
blueprint then the config yang and Module classes can be removed.

There was one complication. The BmpActivator protcol extension is in the
bmp-impl bundle and the service is consumed by the bmp-spi bundle.
The BmpExtensionConsumerContext is provided by the bmp-spi bundle and is
consumed by the bmp-impl bundle. This causes a circular service
dependency which causes a deadlock with blueprint. To avoid this I moved
the BmpActivator and related message and tlv classes (which are only used by
BmpActivator) to a new bundle, bmp-parser-ext. This bundle now provides
the BmpActivator extension provider to the bmp-spi bundle thereby
breaking the circular dependency.

Change-Id: Idbee9a7c2632d1e26c76314213389f83ff8ae5c1
Signed-off-by: Tom Pantelis <tpanteli@brocade.com>