Move pcep base parser Activator to its own bundle
The pcep base parser Activator was moved (and renamed to
BaseParserExtensionActivator) from the impl bundle to a new bundle,
pcep-base-parser, to avoid circular OSGi service dependencies between the
impl and spi bundles when the PCEPDispatcherImpl wiring is cobverted to
blueprint. In addition all the related message et al classes were also
moved to pcep-base-parser.

Change-Id: I9f79bc111743c21e64c6c9444d926c347d392a08
Signed-off-by: Tom Pantelis <tpanteli@brocade.com>