Bug 5061: Introduce BGP RIB deployer + BP wiring
The goal is to migrate RIBImpl configuration from
CSS to BP. The RIBs instances creation needs to be dynamic
and simple - application's compoments wiring should be hidden
from user facing configuration.

The BGP OpenConfig API is an entry point for user's BGP
configuration. The deployer listens to the changes
in a configuration subtree and based on it deploys (starts/restarts/stops) RIB instances.
Instances are registred as OSGi services, where each service is identified by service property.
The service can be used for internal (BGPPeer, ApplicationPeer) and external wiring (topology provider).

Change-Id: I68752ead5b7468c0aada603f52b18e574d4e2cb0
Signed-off-by: Milos Fabian <milfabia@cisco.com>