Update patch set 7
Patch Set 7:

I think withdrawing route is an option in a case of the prefix is not reachable anymore - If I understand it correctly, this not that case.
To avoid the problem described in 6083, the peer connection process (from LocRib perspective) should be like:
1. wait until ODL learns all routes from a new peer
2. best-paths are re-computed
3. only best-paths are advertised to peers (including a new peer)
However, it might not be so easy to implement (not a simple bug fix) and as result a slowdown in a case of many peers are connecting at the same time.

Patch-set: 7