BUG-6069: Fix IAE when POSTing an openconfig neighbor
Try to insert a neighbor will cause an IAE, thrown by
ensureParentsByMerge method. Which force us to write the whole
three since its not able to create mandatory leaf.
As a workaround and for keep functionallity, we make
the leaf as not mandatory and instead we put a default value.

Change-Id: I40ab0b5245bc975db0d56178f879f82e5c5af0dd
Signed-off-by: Claudio <cgaspari@cisco.com>