BUG-6437: Fix race condition when closing singleton group
When Rib Singleton service is closed been the only item on the
service-group. Causes a race condition betwen the removal
of the group, and the new addition to it
(when new Rib/Peer/App instances are added). Making this new
instance never been called by instantiateServiceInstance.
Fix by add an unique an incremental identifier per Rib Instance
created.

Change-Id: I5c1a59bc2d3837b02a5bfa608a98b4c8753b938a
Signed-off-by: Claudio D. Gasparini <cgaspari@cisco.com>