BUG-6437: Fix race condition when closing singleton group
When Rib Singleton service is closed been the only item on the
service-group. Causes a race condition betwen the removal
of the group, and the new addition to it
(when new Rib/Peer/App instances are added). Making this new
instance never been called by instantiateServiceInstance.
Fix by add an unique an incremental identifier per Rib Instance
created.

Change-Id: I48c5219585087ad8ac7647b32120caa30961faab
Signed-off-by: Claudio D. Gasparini <cgaspari@cisco.com>