Update patch set 2
Patch Set 2: Verified-1

Build Failed 

https://jenkins.opendaylight.org/releng/job/bgpcep-verify-carbon-mvn33-openjdk8/119/ : FAILURE

Unable to build due to a non-resolvable parent POM. Search logs for "[FATAL] Non-resolvable parent POM" to find the logs.

The fix for this is usually to make sure your poms are pointing to parents that exist. Double check:

 * groupId is correct and no typos
 * artifactId is correct and no typos
 * version is correct for branch (eg. no Beryllium version inside of Boron branch) ( https://jenkins.opendaylight.org/releng//job/bgpcep-verify-carbon-mvn33-openjdk8/119/ )

https://jenkins.opendaylight.org/releng/job/bgpcep-distribution-check-carbon/74/ : FAILURE

Unable to build due to a non-resolvable parent POM. Search logs for "[FATAL] Non-resolvable parent POM" to find the logs.

The fix for this is usually to make sure your poms are pointing to parents that exist. Double check:

 * groupId is correct and no typos
 * artifactId is correct and no typos
 * version is correct for branch (eg. no Beryllium version inside of Boron branch) ( https://jenkins.opendaylight.org/releng//job/bgpcep-distribution-check-carbon/74/ )

Patch-set: 2