Bug 6585: BGP ChannelOutputLimiter waits forever
After a while (when an output buffer reach upper bound),
BGP's ChannelOutputLimiter handler get stacked waiting for the channel to become writable.
The writability change never happen, however socket flush is invoked,
so the session peer dies on holdtimer expiration.

I was not able to find a root cause of this issue. Different behavior can be cause by
a migration to the native epoll transport. More investigation needed.

Only working soulution I found is calling flush, every time an Update message is written to the output.

Change-Id: I95681d87c30aff82ec360a2d13fc01fe35873507
Signed-off-by: Milos Fabian <milfabia@cisco.com>