Update patch set 1
Patch Set 1:

> Not sure if we need to apply this change in Beryllium, as this
 > problem was not observed there. Or do you have different
 > experience?
 > Also, did you try to tests this particular use-case with Beryllium
 > including this change?
 > Thanks

I tried the scenario mentioned in this bug in Beryllium. Not able to reproduce the hold-timer expiry issue, but what I noticed is without the patch, controller takes a long time (order of minutes) to send all the prefixes to peer. Whereas when the patch is applied, is does it in a much less time (order of seconds). I will email you the pcaps captured for reference

Patch-set: 1