Introduce BGPTableType Registry
The registry is holding supported AFI/SAFI in form of BGPTableType and OpenConfig AfiSafiType.
Currenlty, this two AFI/SAFI representations mapping is held in static map, which lacks of extensibility.
The registry provider context is supposed to be activated in BGP-MP extensions.
The registry would be consumed in BGP configuration mapping procedures.

Change-Id: Ib34147f56d85bf418b29924d6d1f805d4ed1a757
Signed-off-by: Milos Fabian <milfabia@cisco.com>