Update patch set 5
Patch Set 5: Code-Review-1

Does not look a bugfix...Indeed BUG-7102 is not a bug, rather the misconfiguration? (https://wiki.opendaylight.org/view/BGP_LS_PCEP:User_Guide#3-node_Clustering_2).
Anyway, if the goal here is to implement CSS in pcep-topology-provider and programming-impl (BUG-6975), it is not necessary to migrate to blueprint (BUG-6973).

Patch-set: 5