Update patch set 5
Patch Set 5:

> Does not look a bugfix...Indeed BUG-7102 is not a bug, rather the
 > misconfiguration? (https://wiki.opendaylight.org/view/BGP_LS_PCEP:User_Guide#3-node_Clustering_2).
 > Anyway, if the goal here is to implement CSS in pcep-topology-provider
 > and programming-impl (BUG-6975), it is not necessary to migrate to
 > blueprint (BUG-6973).

I agree with you, its not a bug fixing, the enunciate may not be the most appropriate. The intention of this path is to get rid of this errors log on netconf test, and also set some base for BUG-6975 and BUG-6973 without implement them.

Patch-set: 5
Label: Verified=0