BUG-XXX: Conflict Modification
When closing Peer and Rib, can cause some race condition
between the removal of the peer and the rib when updating
DS.
Fix by close peers first and once they are closed, proceed
with peer.
Same its applied in case that peer/rib is updated. First
we close it and then once DS is updated we proceed with
creating the new instance.

Change-Id: I47f10d2518bcabc1aa04222584c6f9086e003185
Signed-off-by: Claudio D. Gasparini <claudio.gasparini@pantheon.tech>