BUG-7976: Race between peer removal and routes update
There is a rece condition when we are advertizing routes,
and a session peer goes down and is removed from DS.
Since advertizement and removal are done by 2 different tx,
we need to do it blocking mode. And not route should be
advertized until peer is removed or peer should not be removed
until routes are advertized.

Change-Id: I46a27871bfa2aa2a632e3bfb76061105c62bb6d2
Signed-off-by: Claudio D. Gasparini <claudio.gasparini@pantheon.tech>