BUG-8156 Terminate PCEP session properly when ServerSessionManager is closed
When the PCEP topology provider is loading the configuration initially,
the ServerSessionManager will be restarted.  Any new PCEP connection
coming in during this period will be rejected by the ServerSessionManager.

This patch is to terminate these rejected PCEP session properly when above
situation happens, so that the session can be established successfully
in the next retry.

Change-Id: Ifa3366c97025a0b31fb7d3ee50b1142f63a6209c
Signed-off-by: Kevin Wang <kevixw@gmail.com>