BUG-8156 Fix PCEP topology registration
It is possible that when a PCEP session is being dropped and being
unregistered from ServerSessionManager, the thread gets switched
thus the topology unregistering cannot finish.

If such situation happens, PCEP session from the same PCEP peer will
not be able to established until after DEADTIMER_EXPIRED, as the
topology instance was not removed correctly.

Change-Id: I647ecfbbb35ef6805d563753b7ebd87edfc350fe
Signed-off-by: Kevin Wang <kevixw@gmail.com>