BUG-8335: Adjust instantiation
The hanging issue stems from the use of a blocking dispatch for DTCL,
which means we are calling into user code while holding listener
registration lock. This does not happen in production environment.

This patch adds an explicit test for this scenario and migrates
the hanging test to call properly initialize its base class.

Also cleans up warnings in the implementation, as it has not seen much
love lately.

Change-Id: I6c424acfad87244c7a23b013ea91da543ecf72a8
Signed-off-by: Claudio D. Gasparini <claudio.gasparini@pantheon.tech>
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>