Bug 8568 - Remove deprecated CSS module tests
Remove Test class that are using deprecated DataProviderService
for Classes
BGPApplicationPeerModule
BGPPeerModule
RIBImplModule
Ipv4ReachabilityTopologyBuilderModule
Ipv6ReachabilityTopologyBuilderModule are deprecated and
will be not tested

Classes AddPathImplModule and BGPTableTypeImplModule are not
deprecated and will be not tested.
Do they need to be deprecated ?

Migrate to odlparent 1.8.0-Carbon Per request of odlparent
project we are downgrading all Nitrogen projects to use the
released odlparent 1.8.0-Carbon to allow for the odlparent
project to start performing semver style releases.
Jira: RELENG-159 RT: 41406

Change-Id: I9b9d3125b6f80ee70c86451f8a46cbd480bde6c2
Signed-off-by: matthieu <mcauffiez@inocybe.com>