Bug 8667 - PCEP: When peer closes got IO exception
When peer closes the pcep session, it results on IO
Exception.
it is due to this close is not recognized as local pcep
PCEP session termination.
It was first implemented as a local pcep session
termination.
https://git.opendaylight.org/gerrit/#/c/906/1/pcep/impl/src/main/java/org/opendaylight/protocol/pcep/impl/PCEPSessionImpl.java
Solution: Use the reason encapsulated in the close message to close the session.

Change-Id: Ia423f5c10b117ef21af5a3bddc28652ebb659e26
Signed-off-by: matcauf <mcauffiez@inocybe.ca>
(cherry picked from commit b8d57a84bd130ff3f98cd08b175bb9da18a6e24e)