Remove stale prerequisite sections
It is generally accepted that ODL requires maven 3.3.9+ in order to
compile correctly.  This cleans up a bunch of places that reference
older minimum versions to avoid confusion.  Instead, we should
recommend that all modules are compiled using the ODL global minimum
in order to promote a common toolset across the project.

Change-Id: I61591cfbd262e712bdd428e8b6697f7c0742dbde
Signed-off-by: Ryan Goulding <ryandgoulding@gmail.com>