Bug 8675: Follow-up fix to new XML parser change in yangtools
Behaviour of the new XML parser in yangtools project needs
to be changed a bit so that RESTCONF/NETCONF projects can be
migrated to the new parser. However, these changes break some
parts of BGPCEP which have already been using it.

This patch provides a fix for the issue and should be merged
right after the following yangtools patch goes in:
https://git.opendaylight.org/gerrit/#/c/59036/

Change-Id: I23a5677e31b01fee7f67c8a5eb6b4081bd976f8b
Signed-off-by: Igor Foltin <igor.foltin@pantheon.tech>