Update patch set 20
Patch Set 20:

It seems there is something wrong with the job.  The job failed due to timeout after 6hrs.  I tried it twice. See
1. https://jenkins.opendaylight.org/releng/job/bgpcep-patch-test-pcep-carbon/1/console
2. https://jenkins.opendaylight.org/releng/job/bgpcep-patch-test-pcep-carbon/2/console

However, the tests started by both jobs actually passed. See
1. https://jenkins.opendaylight.org/releng/job/bgpcep-csit-1node-userfeatures-only-carbon/740/
2. https://jenkins.opendaylight.org/releng/job/bgpcep-csit-1node-userfeatures-only-carbon/743/
3. https://jenkins.opendaylight.org/releng/view/bgpcep/job/bgpcep-csit-1node-periodic-throughpcep-only-carbon/340/
4. https://jenkins.opendaylight.org/releng/view/bgpcep/job/bgpcep-csit-1node-periodic-throughpcep-only-carbon/343/

The job was running slow because it took a extreme long time to assign a VM to run the test.

Patch-set: 20