Bug 9186: BGP session does not get established when send-max flag is not included in configuration
Unlike receive mode, there is no separate flag to specify if send mode should be
advertised in add-path capability. If send-max field is excluded in config,
NPE was being produced and this patch fixes that so that expected add-path
capability can be advertised.
FYI - recent version of OpenConfig model has added a separate flag for send mode
and we should migrate to this new version at some point.
https://github.com/openconfig/public/commit/e020ce4b7e94482a9bbcf8dfb73f3d01dbdda5e6#diff-0d759af0e8ee335a03ba94fd5055d69cL312

Change-Id: I4af1026a153652d888ed678f0fae47c66b636f30
Signed-off-by: Ajay Lele <ajayslele@gmail.com>