BGPCEP-706: Fix BGP Flowspec NumbericOphrand
Right now when serializing a NumbericOphrand, the "end-of-list"
bit is not properly set. The last entry in the ophrand list
should have this bit set according to RFC. Otherwise the packet
will be in malformed format.
This patch fix BGP flowspec NumbericOphrand to set the "end-of-list"
field properly. When the serialization happens, the end-of-list
field will be set in runtime.

Change-Id: Ib8841360b47137dc0ec0f54c3ef156a298ba3c81
Signed-off-by: Kevin Wang <kevixw@gmail.com>